10-kubernetes/workshop/pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.124.2",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
]
